utils restructure
