+++ src/pages/auth/sign-in/index.tsx
@@ -17,7 +17,7 @@ import { handleAxiosError } from '@/api/apiUtils/AxiosConfig';
 import { login } from '@/api/authApi';
 import AuthBase from '@/components/auth/AuthBase';
 import PasswordField from '@/components/auth/PasswordField';
-import { usernameRegex } from '@/utils/utils';
+import { usernameRegex } from '@/utils';
 
 import { UserContext } from '../../../context/UserContext';
